Refactoring

+++ sdk/src/main/java/me/digi/sdk/entities/payload/CredentialsPayload.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class AccessToken(
     @SerializedName("expires_on")
-    val expiresOn: Int = 0,
+    val expiresOn: Long = 0,
     val value: String? = null
 )
 
@@ -12,7 +12,7 @@ data class Identifier(val id: String = "")
 
 data class RefreshToken(
     @SerializedName("expires_on")
-    val expiresOn: Int = 0,
+    val expiresOn: Long = 0,
     val value: String? = null
 )
